Backend Base Structure
Created the basic structure of the backend API, with an example module to show how it would work. The example module will be removed for production release. Updated backend README to include library/package installations and the command to run the Flask application.  Added a .gitignore file so we can ignore files/folders that dont need to be on the repo, and for this case, pycache files auto generated by Python after executing the scripts for the first time.